+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Keep a local seismic catalog.
 
 Copyright (c) 2022-2023 Claudio Satriano <satriano@ipgp.fr>
 
+## unreleased
+
+- `--format` option for `seiscat print`: current possibilities are `table`
+  (default) and `stats`
+
 ## v0.4 - 2023-06-20
 
 - Command line autocompletion, thanks to argcomplete